additional lint cleanup
