+++ build-system/compile/compile.js
@@ -42,7 +42,7 @@ const MAX_PARALLEL_CLOSURE_INVOCATIONS = argv.single_pass ? 1 : 4;
  * Prefixes the the tmp directory if we need to shadow files that have been
  * preprocess by babel in the `dist` task.
  *
- * @param {!Array<string>}
+ * @param {!Array<string>} paths
  * @return {!Array<string>}
  */
 function convertPathsToTmpRoot(paths) {
